refactor(reports): move BenchmarkComparison to shared types
